backend/pyproject.toml
[project]
name = "pycarrotquant"
version = "0.0.0"
description = "Carrot Quant IDE Backend"
requires-python = ">=3.14"
dependencies = [
    "akshare>=1.18.21",
    "apscheduler>=3.11.2",
    "backend>=0.2.4.1",
    "duckdb>=1.4.4",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "pandas>=3.0.0",
    "pandas-stubs==2.3.3.260113",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.5",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
]
